py/binary: Make return type of mp_binary_get_size size_t instead of int.
Fixes sign-compare warning.

+++ py/binary.h
@@ -32,7 +32,7 @@
 // (underlyingly they're same).
 #define BYTEARRAY_TYPECODE 0
 
-int mp_binary_get_size(char struct_type, char val_type, mp_uint_t *palign);
+size_t mp_binary_get_size(char struct_type, char val_type, mp_uint_t *palign);
 mp_obj_t mp_binary_get_val_array(char typecode, void *p, mp_uint_t index);
 void mp_binary_set_val_array(char typecode, void *p, mp_uint_t index, mp_obj_t val_in);
 void mp_binary_set_val_array_from_int(char typecode, void *p, mp_uint_t index, mp_int_t val);
